Score is now part of game class
